fix(omnihive): build manual commands test folder crash
The build manual command now ignores the delete test folder step if the tests folder does not exist.

+++ buildManual.ts
@@ -199,7 +199,9 @@ const getPublishFolders = () => {
 };
 
 const publish = (directory: string) => {
-    fse.rmdirSync(path.join(directory, `tests`), { recursive: true });
+    if (fse.existsSync(path.join(directory, "tests"))) {
+        fse.rmdirSync(path.join(directory, `tests`), { recursive: true });
+    }
 
     let publishString: string = "npm publish";
